+++ modules/try_split/native/TrySplitNoCriteria.h
@@ -15,5 +15,5 @@ public:
 
     virtual TrySplitCriteriaI* Clone() const;
 
-    virtual bool TrySplit(int depth, int numberOfDatapoints) const;
+    virtual bool TrySplit(int depth, double numberOfDatapoints) const;
 };
